Gemini_whatsapp_Chatbot/requirements.txt
flask
twilio
langchain
langchain-community
pypdf
python-dotenv
google-generativeai
langchain_google_genai
langchain-chroma
